+++ kinto/core/testing.py
@@ -8,9 +8,10 @@ import webtest
 from cornice import errors as cornice_errors
 from pyramid.url import parse_url_overrides
 
-from kinto.core import DEFAULT_SETTINGS, statsd
+from kinto.core import DEFAULT_SETTINGS
 from kinto.core.storage import generators
 from kinto.core.utils import encode64, follow_subrequest, memcache, sqlalchemy
+from kinto.plugins import statsd
 
 
 skip_if_ci = unittest.skipIf("CI" in os.environ, "ci")
