+++ tests/plugins/test_history.py
@@ -22,7 +22,7 @@ class PluginSetup(unittest.TestCase):
             "includes": "kinto.plugins.history",
         }
         config = testing.setUp(settings=settings)
-        with mock.patch("kinto.core.statsd.Client.timer") as mocked:
+        with mock.patch("kinto.plugins.statsd.Client.timer") as mocked:
             kinto_main(None, config=config)
             mocked.assert_called_with("plugins.history")
